initial addq and q support
/addq first and forward then? or the other way around? which is more
convenient? now it's /addq first because the user is expected to be
looking at that chat already.

The quote files require text and message ids and from-ids and stuff so
just pickle them and think later if that's a good thing. Wrap quotes in
a TgQuote tuple to be able to still use "searchterm in quote".